rearranged cli.go
